ported our interpolators from previous processing code & added convenience method to switch to different interpolator settings.

+++ openframeworks/wiretouch_monitor/src/interpolator2x2.h
@@ -18,7 +18,7 @@ class wtmInterpolator2x2 : public wtmInterpolator {
     protected:
         double a, b, c, d;
     
-        virtual double interp_func(double a, double b, float f);
+        virtual double interp_func(double a, double b, double f);
         double interpolate4(int x, int y, int ix, int iy, float fx, float fy);
         void beginInterpolate4(int x, int y);
 };
